Reduce common code in serai-coordinator-tests with amore involved new_test
